Don't send OSC packets back to the source and don't add source as remote
This prevents network loops. When sending a 'button pressed' OSC message
QLC+ will reply back. That reply socket address would be added to the
remotes (because different ports) and we would send new messages to it.
When 'button pressed' and 'button depressed' are send in rapid
succession, we would receive the 'button pressed' reply from QLC+ later
than we send 'button depressed'.